aider: Updated image path in markdown file to be relative.
# Aider chat conversation:

USER: what if Silly Drawing.jpg is in the same folder

ASSISTANT: If "Silly Drawing.jpg" is in the same folder as the "WGIP.md" file, then the image link in the markdown file is correct as it is. The syntax `![Silly Drawing](Silly Drawing.jpg)` will display the image in the markdown file if the image is located in the same directory as the markdown file.

However, if you want to ensure that the image is displayed correctly regardless of the directory structure, you could use a relative path from the markdown file to the image file. If the image file is in the same directory as the markdown file, the relative path would be `./Silly Drawing.jpg`.

Here is the *SEARCH/REPLACE* block for this change:

WGIP.md
<source>
<<<<<<< SEARCH
![Silly Drawing](Silly Drawing.jpg)
=======
![Silly Drawing](./Silly Drawing.jpg)
>>>>>>> REPLACE
</source>